+++ src/main/java/jp/igapyon/cityinfojp/disp/DisplayCityInfoEntry.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package jp.igapyon.cityinfojp.dyn;
+package jp.igapyon.cityinfojp.disp;
 
 import java.util.List;
 
 /**
- * HTMLレンダリング用のエントリ
+ * cityinfo の HTML レンダリング用のエントリ
  * 
  * @author Toshiki Iga
  */
