+++ src/main/java/jp/igapyon/cityinfojp/disp/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 画面表示または静的HTML生成で利用する機能。
+ */
+package jp.igapyon.cityinfojp.disp;
