+++ src/main/resources/static/about.html
@@ -181,7 +181,7 @@
     </main>
 
     <footer class="container">
-        <p>&copy; cityinfojp 2020 : 2020-04-18 23:31</p>
+        <p>&copy; cityinfojp 2020 : 2020-04-18 23:58</p>
     </footer>
 
     <!-- Bootstrap JavaScript -->
